Add Settings for easier item attribute creation
