+++ src/main/java/io/github/cleanroommc/millennium/common/blocks/MillenniumBlocks.java
@@ -1,0 +1,4 @@
+package io.github.cleanroommc.millennium.common.blocks;
+
+public class MillenniumBlocks {
+}
